requirements.txt
annotated-types==0.6.0
anyio==4.1.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
Django==5.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
openai
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
wcwidth==0.2.12
whitenoise==6.6.0
gunicorn==20.0.4
openai
langchain
streamlit
langchain-openai
unstructured
tiktoken
pinecone-client
pypdf
openai
huggingface_hub